--- __main__.py
@@ -1,3 +1,0 @@
-import uvicorn
-
-uvicorn.run("app:app", port=9999, reload=True)
