+++ docker-compose.yml
@@ -1,8 +1,8 @@
 version: '3.9'
 
 services:
-  api-db:
-    image: postgres:15-bullseye
+  database:
+    image: postgres:bullseye
     ports:
       - "5432:5432"
     environment:
