+++ src/dbdomain/__init__.py
@@ -1,7 +1,7 @@
 host="localhost:5432"
 TORTOISE_ORM_CONFIG:dict = {
 "connections":{
-    "dev-string-connection":f"postgres://postgres:postgres@{host}/postgres"
+    "dev-string-connection":f"postgres://postgres:postgres@localhost:5432/postgres"
 },
 "apps":{
     "commands-api":{
