+++ src/prisma/schema.prisma
@@ -1,0 +1,37 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+generator client {
+  provider             = "prisma-client-py"
+  interface            = "asyncio"
+  recursive_type_depth = 5
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model User {
+  id Int @id @default(autoincrement())
+
+  email String @unique
+
+  name String?
+
+  posts Post[]
+}
+
+model Post {
+  id Int @id @default(autoincrement())
+
+  title String
+
+  content String?
+
+  published Boolean @default(false)
+
+  author User @relation(fields: [authorId], references: [id])
+
+  authorId Int @unique
+}
